requirements.txt
requests
selenium
bs4
python-dotenv